fix compiler warning
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/aims/aims-free/trunk@82021 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ aimsalgo/src/aimsalgo/transform/least_square_estimation.cc
@@ -314,10 +314,10 @@ TranslationLeastSquareEstimation::computeTranslationMotion()
 {
   if( _pointsFrom.size() != _pointsTo.size() )
     throw runtime_error("TranslationLeastSquareEstimation:Both vectors must have same size !") ;
-  unsigned size = _pointsFrom.size();
+  size_t size = _pointsFrom.size();
 
   _motion->translation() = Point3df( 0., 0., 0. ) ;
-  for( int n = 0 ; n < size ; ++n ){
+  for( size_t n = 0 ; n < size ; ++n ){
     _motion->translation() += _pointsTo[n] - _pointsFrom[n] ;
   }
   _motion->translation() /= float(size) ;
